test: modify array test code and add bug test case
